test(store): isolate future-schema preservation coverage
